+++ ng-workspace/projects/OpenWaterFoundation/common/leaflet/map/map.module.ts
@@ -3,6 +3,7 @@ import { CommonModule }                from '@angular/common';
 import { BrowserAnimationsModule }     from '@angular/platform-browser/animations';
 import { BrowserModule }               from '@angular/platform-browser';
 
+import { MatSlideToggleModule }        from '@angular/material/slide-toggle';
 import { MatTooltipModule }            from '@angular/material/tooltip';
 // The shared pipes secondary entry point for the common library.
 import { PipesModule }                 from '@OpenWaterFoundation/common/pipes';
@@ -24,6 +25,7 @@ import { LegendLayerGroupModule }      from './legend/legend-layer-group/legend-
     BrowserAnimationsModule,
     BrowserModule,
 
+    MatSlideToggleModule,
     MatTooltipModule,
 
     PipesModule,