+++ ng-workspace/projects/OpenWaterFoundation/common/pipes/pipes.module.ts
@@ -4,17 +4,20 @@ import { CommonModule }      from '@angular/common';
 import { JustificationPipe } from './justification.pipe';
 import { MenuDisablePipe }   from './menu-disable.pipe';
 import { ZoomDisablePipe }   from './zoom-disable.pipe';
+import { SymbolCheckPipe } from './symbol-check.pipe';
 
 
 @NgModule({
   declarations: [
     JustificationPipe,
     MenuDisablePipe,
-    ZoomDisablePipe
+    ZoomDisablePipe,
+    SymbolCheckPipe
   ],
   exports: [
     JustificationPipe,
     MenuDisablePipe,
+    SymbolCheckPipe,
     ZoomDisablePipe
   ],
   imports: [