+++ ng-workspace/projects/OpenWaterFoundation/common/pipes/public-api.ts
@@ -1,4 +1,5 @@
 export * from './pipes.module';
 export * from './justification.pipe';
 export * from './menu-disable.pipe';
+export * from './symbol-check.pipe';
 export * from './zoom-disable.pipe';